Fix architecture second time
