+++ api/user.go
@@ -21,11 +21,11 @@ type UserAPI struct {
 	Tx *gmtoken.GmtokenTx
 }
 
-type CreateUserRes struct {
+type CreateUserResponse struct {
 	Token string `json:"token"`
 }
 
-type GetUserRes struct {
+type GetUserResponse struct {
 	Name           string `json:"name"`
 	Address        string `json:"address"`
 	GmtokenBalance int    `json:"gmtoken_balance"`
@@ -38,43 +38,43 @@ type GetUserRes struct {
 func (a *UserAPI) CreateUser(ctx *gin.Context) {
 	body, err := ioutil.ReadAll(ctx.Request.Body)
 	defer ctx.Request.Body.Close()
-	if success := util.SuccessOrAbort(ctx, http.StatusBadRequest, err); !success {
+	if success := successOrAbort(ctx, http.StatusBadRequest, err); !success {
 		return
 	}
 	var user model.User
-	if success := util.SuccessOrAbort(ctx, http.StatusBadRequest, json.Unmarshal(body, &user)); !success {
+	if success := successOrAbort(ctx, http.StatusBadRequest, json.Unmarshal(body, &user)); !success {
 		return
 	}
 	userId, err := util.CreateUUId()
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, err); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, err); !success {
 		return
 	}
 	user.UserID = userId
 	// 新規ユーザの秘密鍵を生成
 	privateKeyHex, err := util.CreatePrivateKey()
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, err); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, err); !success {
 		return
 	}
 	user.PrivateKey = privateKeyHex
 	// ゲームトークンを100だけ鋳造し、新規ユーザに付与
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, a.Tx.MintGmtoken(100, user.PrivateKey)); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, a.Tx.MintGmtoken(100, user.PrivateKey)); !success {
 		return
 	}
 	//	INSERT INTO `users` (`user_id`,`name`,`private_key`)
 	//	VALUES ('95daec2b-287c-4358-ba6f-5c29e1c3cbdf','aaa','6e7eada90afb7e84bf5b4498c6adaa2d4014904644637d5fb355266944fbf93a')
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, a.DB.CreateUser(user)); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, a.DB.CreateUser(user)); !success {
 		return
 	}
 	// ユーザIDの文字列からjwtでトークン作成
 	token, err := a.Auth.CreateToken(userId)
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, err); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, err); !success {
 		return
 	}
 	// token = "生成されたトークンの文字列"
-	createUserRes := &CreateUserRes{
+	createUserResponse := &CreateUserResponse{
 		Token: token,
 	}
-	ctx.JSON(http.StatusOK, createUserRes)
+	ctx.JSON(http.StatusOK, createUserResponse)
 	// {"token":"生成されたトークンの文字列"}が返る
 }
 
@@ -84,24 +84,24 @@ func (a *UserAPI) CreateUser(ctx *gin.Context) {
 // コントラクトからそのユーザアドレスのゲームトークン残高を取り出し、返す
 func (a *UserAPI) GetUser(ctx *gin.Context) {
 	userId, err := a.Auth.GetUserId(ctx)
-	if success := util.SuccessOrAbort(ctx, http.StatusBadRequest, err); !success {
+	if success := successOrAbort(ctx, http.StatusBadRequest, err); !success {
 		return
 	}
 	// SELECT * FROM `users` WHERE user_id = '95daec2b-287c-4358-ba6f-5c29e1c3cbdf'
 	user, err := a.DB.GetUser(userId)
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, err); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, err); !success {
 		return
 	}
 	address, balance, err := a.Tx.GetAddressBalance(user.PrivateKey)
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, err); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, err); !success {
 		return
 	}
-	getUserRes := &GetUserRes{
+	getUserResponse := &GetUserResponse{
 		Name:           user.Name,
 		Address:        address.String(),
 		GmtokenBalance: balance,
 	}
-	ctx.JSON(http.StatusOK, getUserRes)
+	ctx.JSON(http.StatusOK, getUserResponse)
 	// {"name":"aaa","address":"0x7a242084216fC7810aAe02c6deE5D9092C6B8fb9","gmtoken_balance":40}が返る
 	// 有効期限が切れると{"code":400,"message":"Token is expired"}が返る
 }
@@ -111,21 +111,21 @@ func (a *UserAPI) GetUser(ctx *gin.Context) {
 // トークンからユーザIDを取り出し、dbからそのユーザIDのユーザの情報を更新
 func (a *UserAPI) UpdateUser(ctx *gin.Context) {
 	userId, err := a.Auth.GetUserId(ctx)
-	if success := util.SuccessOrAbort(ctx, http.StatusBadRequest, err); !success {
+	if success := successOrAbort(ctx, http.StatusBadRequest, err); !success {
 		return
 	}
 	body, err := ioutil.ReadAll(ctx.Request.Body)
 	defer ctx.Request.Body.Close()
-	if success := util.SuccessOrAbort(ctx, http.StatusBadRequest, err); !success {
+	if success := successOrAbort(ctx, http.StatusBadRequest, err); !success {
 		return
 	}
 	var user model.User
-	if success := util.SuccessOrAbort(ctx, http.StatusBadRequest, json.Unmarshal(body, &user)); !success {
+	if success := successOrAbort(ctx, http.StatusBadRequest, json.Unmarshal(body, &user)); !success {
 		return
 	}
 	// dbでnameとaddressを更新
 	// UPDATE `users` SET `name`='bbb' WHERE user_id = '95daec2b-287c-4358-ba6f-5c29e1c3cbdf'
-	if success := util.SuccessOrAbort(ctx, http.StatusInternalServerError, a.DB.UpdateUser(user, userId)); !success {
+	if success := successOrAbort(ctx, http.StatusInternalServerError, a.DB.UpdateUser(user, userId)); !success {
 		return
 	}
 	ctx.JSON(http.StatusOK, nil)
