+++ router/router.go
@@ -5,20 +5,14 @@ import (
 
 	"game-api-gin/api"
 	"game-api-gin/auth"
-	"game-api-gin/config"
 	"game-api-gin/database"
 	"game-api-gin/gmtoken"
 
 	"github.com/gin-gonic/gin"
 )
 
-func CreateRouter(db *database.GormDatabase, config *config.Config) (*gin.Engine, error) {
+func CreateRouter(auth *auth.Auth, db *database.GormDatabase, gmtokenTx *gmtoken.GmtokenTx) *gin.Engine {
 	router := gin.Default()
-	auth := auth.NewAuth(config)
-	gmtokenTx, err := gmtoken.NewGmtokenTx(config)
-	if err != nil {
-		return nil, err
-	}
 	userHandler := &api.UserAPI{
 		Auth: auth,
 		DB: db,
@@ -39,7 +33,7 @@ func CreateRouter(db *database.GormDatabase, config *config.Config) (*gin.Engine
 	router.PUT("/user/update", userHandler.UpdateUser)
 	router.POST("/gacha/draw", gachaHandler.DrawGacha)
 	router.GET("/character/list", characterHandler.GetCharacterList)
-	return router, nil
+	return router
 }
 
 // {"message":"Hello World"}をlocalhost:8080画面に表示
